Dialog / DialogBot, nitfixes; this commit dedicated to @technillogue
